+++ src/mnemonic.hpp
@@ -9,10 +9,11 @@
 #include "mnemonic_bit.hpp"
 #include "mnemonic_calc.hpp"
 #include "mnemonic_data.hpp"
+#include "mnemonic_dec.hpp"
 #include "mnemonic_djnz.hpp"
 #include "mnemonic_ex.hpp"
 #include "mnemonic_im.hpp"
-#include "mnemonic_increment.hpp"
+#include "mnemonic_inc.hpp"
 #include "mnemonic_io.hpp"
 #include "mnemonic_jp.hpp"
 #include "mnemonic_jr.hpp"
